requirements-prod.txt
pandas>=2.2
requests>=2.31
python-dotenv>=1.0
beautifulsoup4>=4.12
scikit-learn>=1.4
boto3>=1.34
pyarrow>=15.0
